BRANCH Instruction
This is based in version 1, now program counter is separated from the core registers because i was having lots of syncronization issues, now its good. However there are some timing critical warnings when i ran the generate bitstream and the implementation, i dont know how to correct them but when i programed the fpga the output and all was fine.